# swagger.yaml
swagger: '2.0'
info:
  description: Findfreelancer api
  version: 1.0.0
  title: Findfreelancer API
consumes:
  - application/json
paths:
...
  /order:
    post:
      tags:
        - order
      summary: Создать заказ
      description: Создание заказа
      parameters:
        - name: order
          in: body
          description: Заказ
          schema:
            $ref: '#/definitions/Order'
      responses:
        201:
          description: Created
          schema:
            $ref: '#/definitions/Order'
    get:
      tags:
        - order
      summary: Получить все заказы
      description: Получить все актуальные заказы
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/Orders'
...
  /vacancy:
    post:
      tags:
        - vacancy
      summary: Создать вакансию
      description: Создать вакансию
      parameters:
        - name: vacancy
          in: body
          description: информацию о вакансии
          required: true
          schema:
            $ref: '#/definitions/Vacancy'
      responses:
        201:
          description: Created
          schema:
            $ref: '#/definitions/Vacancy'
    get:
      tags:
        - vacancy
      summary: Получить все вакансии
      description: Получить все актуальные вакансии
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/Vacancies'
...
definitions:
...
  Order:
    type: object
    properties:
      id:
        type: number
        example: 228
        readOnly: true
      order_name:
        type: string
        example: "Сверстать сайт"
      category:
        type: string
        example: "Back"
      customer_id:
        type: number
        example: 322
      executor_id:
        type: number
        example: 42
        readOnly: true
      deadline:
        type: number
        example: 1617004533
        format: Unix-Time
      budget:
        type: number
        example: 1488
      desctiption:
        type: string
        example: "Pomogite sdelat API"
      user_login:
        type: string
        example: "AlexDarkStalker98"
        readOnly: true
      user_img:
        type: string
        example: "findreelancer.ru/images/kek.jpg"
        readOnly: true
...
  Orders:
    type: array
    items:
      $ref: '#/definitions/Order'
...
  Vacancy:
    type: object
    properties:
      id:
        type: number
        example: 228
        readOnly: true
      customer_id:
        type: number
        format: uint64
        example: 228
      category:
        type: string
        example: "Back"
      vacancy_name:
        type: string
        example: "Junior фронтенд разработчик"
      description:
        type: string
        example: "В супер компанию требуется джун фронтенд разработчик"
      salary:
        type: number
        example: 1488228
      user_login:
        type: string
        example: "AlexMem228"
        readOnly: true
      user_img:
        type: string
        example: "findreelancer.ru/images/kek.jpg"
        readOnly: true
  Vacancies:
    type: array
    items:
      $ref: '#/definitions/Vacancy'
...
host: findfreelancer.ru
basePath: /
schemes:
  - https